+++ src/framework/Localization.cpp
@@ -59,6 +59,7 @@ static const wchar_t* const table[kLangCount][kStrCount] = {
     // Overlay
     L"F1: Settings",
     L" (%.0f%% in use)",
+    L" (%d%% in use)",
 
     // System info
     L"Unknown CPU",
@@ -115,6 +116,7 @@ static const wchar_t* const table[kLangCount][kStrCount] = {
     // Overlay
     L"F1: \xC124\xC815\xCC3D",
     L" (%.0f%% \xC0AC\xC6A9 \xC911)",
+    L" (%d%% \xC0AC\xC6A9 \xC911)",
 
     // System
     L"\xC54C \xC218 \xC5C6\xB294 CPU",
@@ -171,6 +173,7 @@ static const wchar_t* const table[kLangCount][kStrCount] = {
     // Overlay
     L"F1: \u8A2D\u5B9A",
     L" (%.0f%% \u4F7F\u7528\u4E2D)",
+    L" (%d%% \u4F7F\u7528\u4E2D)",
 
     // System
     L"\u4E0D\u660E\u306ACPU",
@@ -227,6 +230,7 @@ static const wchar_t* const table[kLangCount][kStrCount] = {
     // Overlay
     L"F1: \u8BBE\u7F6E",
     L" (%.0f%% \u4F7F\u7528\u4E2D)",
+    L" (%d%% \u4F7F\u7528\u4E2D)",
 
     // System
     L"\u672A\u77E5 CPU",
@@ -283,6 +287,7 @@ static const wchar_t* const table[kLangCount][kStrCount] = {
     // Overlay
     L"F1: Einstellungen",
     L" (%.0f%% aktiv)",
+    L" (%d%% aktiv)",
 
     // System
     L"Unbekannte CPU",
@@ -339,6 +344,7 @@ static const wchar_t* const table[kLangCount][kStrCount] = {
     // Overlay
     L"F1: Param\u00E8tres",
     L" (%.0f%% utilis\u00E9)",
+    L" (%d%% utilis\u00E9)",
 
     // System
     L"CPU inconnu",
@@ -395,6 +401,7 @@ static const wchar_t* const table[kLangCount][kStrCount] = {
     // Overlay
     L"F1: Configuraci\u00F3n",
     L" (%.0f%% en uso)",
+    L" (%d%% en uso)",
 
     // System
     L"CPU desconocida",